Change all mod.go but ones inside message
