+++ be1-go/channel/authentication/authentication_test.go
