+++ be1-go/channel/coin/coin_test.go
