+++ be1-go/channel/consensus/consensus.go
