+++ be1-go/channel/reaction/reaction_test.go
