+++ be1-go/cli/cli_test.go
