+++ be1-go/internal/depgraph/depgraph.go
